pass font_size on update response; change theme, fontsize, lang on front

+++ server/users/internal/handler/http/update_user.go
@@ -42,11 +42,11 @@ func (h *Handler) Update(w http.ResponseWriter, req *http.Request) (err error) {
 
 	fontSize, err := strconv.Atoi(newFontSize)
 	if err != nil {
-		// TODO: log error
+		// left font size as is
 		fontSize = 0
 	}
 
-	err = h.controller.UpdateUser(req.Context(), user.ID, newLogin, newTheme, newLang, fontSize)
+	err = h.controller.UpdateUser(req.Context(), user.ID, newLogin, newTheme, newLang, int32(fontSize))
 	if err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(server.ServerResponse{
